1. Nodes module decomposition. 2. Tests and other files changed appropriately. 3. Small fix in pyproject.toml.
