+++ pyproject.toml
@@ -8,8 +8,8 @@ authors = ["Roman Netrogolov <romius2001@gmail.com>",
 	   "Nikita Kovalev Yasen <Nikitoskova123@gmail.com>",
 	   "Anna Bubnova <banuba313@gmail.com>",
 	   "Yury Kaminsky <jkaminski@niuitmo.ru>"]
-license = "MIT"
-readme = "readme.md"
+license = "BSD-3-Clause"
+readme = "readme.rst"
 repository = "https://github.com/ITMO-NSS-team/BAMT"
 
 classifiers = ["Programming Language :: Python :: 3",
